Implement LED light duration functions

+++ naraapi.h
@@ -232,8 +232,8 @@ namespace Nara
 		void SetLightSpeed(int key, int fn, LED_Speed speed);
 		LED_Speed GetLightSpeed(int key, int fn);
 
-		void SetLightDuration(int key, int fn, int duration);
-		int GetLightDuration(int key, int fn);
+		void SetLightDuration(int key, int fn, uint8_t duration);
+		uint8_t GetLightDuration(int key, int fn);
 	};
 
 	// Basically wraps hid_init() to a T. Returns 0 on success, -1 otherwise.
